requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
matplotlib>=3.5.0
albumentations>=1.3.0
timm>=0.6.0
einops>=0.6.0
tqdm>=4.64.0
tensorboard>=2.9.0
scikit-learn>=1.1.0
lmdb>=1.3.0
six>=1.16.0